requirements.txt
google-cloud-aiplatform
google-cloud-logging>=3.8.0
google-cloud-firestore
streamlit
google-genai
pandas>=1.3.0
matplotlib>=3.4.0
numpy>=1.20.0